make leasing a shared value => clone
This means that `our leased` is a very general
default that works well whenever you are reading
things. Basically "our leased" means "something I can
read for as long as this place is alive".

This helps to align the left-to-right evaluation
semantics in tricky situations, as well, since reserving
a location that is shared can always just clone it.